serde_wasm_bindgen

Struct Serializer

pub struct Serializer { /* private fields */ }

A serde::Serializer that converts supported Rust values into a JsValue.

Implementations

impl Serializer

pub const fn new() -> Self

Creates a new default Serializer.

pub const fn json_compatible() -> Self

Creates a JSON compatible serializer. This uses null instead of undefined, and uses plain objects instead of ES maps. So you will get the same result of JsValue::from_serde, and you can stringify results to JSON and store it without data loss.

pub const fn serialize_missing_as_null(self, value: bool) -> Self

Set to true to serialize (), unit structs and Option::None to null instead of undefined in JS. false by default.

pub const fn serialize_maps_as_objects(self, value: bool) -> Self

Set to true to serialize maps into plain JavaScript objects instead of ES2015 Maps. false by default.

pub const fn serialize_large_number_types_as_bigints(self, value: bool) -> Self

Set to true to serialize 64-bit numbers to JavaScript BigInt instead of plain numbers. false by default.

pub const fn serialize_bytes_as_arrays(self, value: bool) -> Self

Set to true to serialize bytes into plain JavaScript arrays instead of ES2015 Uint8Arrays. false by default.